requirements.txt
aniso8601==8.0.0
certifi==2020.4.5.1
chardet==3.0.4
Click==7.0
ffmpeg-python==0.2.0
Flask==1.1.1
Flask-RESTful==0.3.7
future==0.18.2
gunicorn==20.0.4
idna==2.9
imutils==0.5.3
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
numpy==1.18.1
opencv-python==4.1.2.30
Pillow==7.0.0
pytesseract==0.3.1
pytz==2019.3
requests==2.23.0
six==1.14.0
urllib3==1.25.9
Werkzeug==0.16.0
